requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Werkzeug==3.1.3
email-validator==2.3.0
WTForms==3.2.1
requests==2.32.5
python-dotenv==1.2.1